Access OpenVPN remotes from separate destination
