+++ Packages/App/Sources/AppUIMain/Views/Modules/OpenVPNView+Remotes.swift
@@ -23,10 +23,76 @@
 //  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
 //
 
+import CommonUtils
+import PassepartoutKit
 import SwiftUI
 
 extension OpenVPNView {
     struct RemotesView: View {
+        let endpoints: [ExtendedEndpoint]
+
+        @ObservedObject
+        var excludedEndpoints: ObservableList<ExtendedEndpoint>
+
+        let remotesRoute: (any Hashable)?
+
+        var body: some View {
+            Form {
+                ForEach(endpoints, id: \.rawValue) { remote in
+                    SelectableRemoteButton(
+                        remote: remote,
+                        all: Set(endpoints),
+                        excludedEndpoints: excludedEndpoints
+                    )
+                }
+                if let remotesRoute {
+                    NavigationLink(Strings.Global.Actions.edit, value: remotesRoute)
+                }
+            }
+            .themeForm()
+            .navigationTitle(Strings.Modules.Openvpn.remotes)
+        }
+    }
+}
+
+private struct SelectableRemoteButton: View {
+    let remote: ExtendedEndpoint
+
+    let all: Set<ExtendedEndpoint>
+
+    @ObservedObject
+    var excludedEndpoints: ObservableList<ExtendedEndpoint>
+
+    var body: some View {
+        Button {
+            if excludedEndpoints.contains(remote) {
+                excludedEndpoints.remove(remote)
+            } else {
+                if remaining.count > 1 {
+                    excludedEndpoints.add(remote)
+                }
+            }
+        } label: {
+            HStack {
+                EndpointCardView(endpoint: remote)
+                Spacer()
+                ThemeImage(.marked)
+                    .opaque(!excludedEndpoints.contains(remote))
+            }
+            .contentShape(.rect)
+        }
+        .buttonStyle(.plain)
+    }
+
+    private var remaining: Set<ExtendedEndpoint> {
+        all.filter {
+            !excludedEndpoints.contains($0)
+        }
+    }
+}
+
+extension OpenVPNView {
+    struct EditableRemotesView: View {
 
         @EnvironmentObject
         private var theme: Theme
@@ -37,7 +103,7 @@ extension OpenVPNView {
         var body: some View {
             Form {
                 theme.listSection(
-                    Strings.Modules.Openvpn.remotes,
+                    nil,
                     addTitle: Strings.Global.Actions.add,
                     originalItems: $remotes,
                     itemLabel: {
@@ -51,6 +117,7 @@ extension OpenVPNView {
             }
             .labelsHidden()
             .themeForm()
+            .navigationTitle(Strings.Modules.Openvpn.remotes)
         }
     }
 }
