+++ Packages/App/Sources/UILibrary/Theme/UI/Theme+Views.swift
@@ -31,7 +31,7 @@ import SwiftUI
 #if !os(tvOS)
 extension Theme {
     public func listSection<ItemView: View, T: EditableValue>(
-        _ title: String,
+        _ title: String?,
         addTitle: String,
         originalItems: Binding<[T]>,
         emptyValue: (() async -> T)? = nil,
